git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/branches/ogl-es@5416 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CCSMLoader.cpp
@@ -50,7 +50,7 @@ namespace scene
 	{
 		BinaryFileReader(io::IReadFile* pFile) : file(pFile) { }
 
-		s32 readBuffer(void* buffer, s32 len)
+		size_t readBuffer(void* buffer, s32 len)
 		{
 			return file->read(buffer,len);
 		}
